+++ lib/screens/login/login.dart
@@ -2,6 +2,7 @@ import 'package:animated_collection/animated_collection.dart';
 import 'package:firebase_analytics/firebase_analytics.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_svg/flutter_svg.dart';
 import 'package:go_router/go_router.dart';
 import 'package:logging/logging.dart';
 
@@ -89,6 +90,10 @@ class __LoginScreenContentState extends State<_LoginScreenContent> {
       key: _formKey,
       child: ListView(
         children: <Widget>[
+          SvgPicture.asset(
+            'assets/logo.svg',
+            height: 120,
+          ),
           Padding(
             padding: const EdgeInsets.symmetric(horizontal: 16.0),
             child: TextFormField(